# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "tech-software-engineer-consultant-int-backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/BonDeLivraison/Create-BonDeLivraison": {
      "post": {
        "tags": [
          "BonDeLivraison"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BLCreateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
...
    "/api/BonDeLivraison/get-Bon-De-Livraison-ID/{id}": {
      "get": {
        "tags": [
          "BonDeLivraison"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BLDTO"
            }
          }
        }
      }
    },
    "/api/BonDeLivraison/get-Bon-De-Livraison-Ref/{RefBL}": {
      "get": {
        "tags": [
          "BonDeLivraison"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "RefBL",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BLDTO"
            }
          }
        }
      }
    },
    "/api/BonDeLivraison/update-BonDeLivraison/{id}": {
      "put": {
        "tags": [
          "BonDeLivraison"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "bonDeLivraisonId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BLUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/api/BonDeLivraison/delete-BonDeLivraison/{id}": {
      "delete": {
        "tags": [
          "BonDeLivraison"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/api/BonDeSortie/Create-Bon-De-Sortie": {
      "post": {
        "tags": [
          "BonDeSortie"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BonDeSortieCreateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
...
    "/api/BonDeSortie/get-Bon-De-Sortie-ByID/{id}": {
      "get": {
        "tags": [
          "BonDeSortie"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BonDeSortieDTO"
            }
          }
        }
      }
    },
    "/api/BonDeSortie/get-Bon-De-Sortie-ByRef/{reference}": {
      "get": {
        "tags": [
          "BonDeSortie"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "reference",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BonDeSortieDTO"
            }
          }
        }
      }
    },
    "/api/BonDeSortie/update-BonDeSortie/{id}": {
      "put": {
        "tags": [
          "BonDeSortie"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "bonDeSortieId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BonDeSortieUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/api/BonDeSortie/delete-BonDeSortie/{id}": {
      "delete": {
        "tags": [
          "BonDeSortie"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BLCreateDTO": {
      "type": "object",
      "properties": {
        "titreClient": {
          "$ref": "#/definitions/Titre"
        },
        "nomClient": {
          "type": "string"
        },
        "adresseClient": {
          "type": "string"
        },
        "mfClient": {
          "type": "string"
        },
        "gsmClient": {
          "format": "int64",
          "type": "integer"
        },
        "proprietaireId": {
          "format": "int32",
          "type": "integer"
        },
        "commandes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Commande"
          }
        },
        "remiseBL": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "BLDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "titreClient": {
          "$ref": "#/definitions/Titre"
        },
        "nomClient": {
          "type": "string"
        },
        "adresseClient": {
          "type": "string"
        },
        "mfClient": {
          "type": "string"
        },
        "gsmClient": {
          "format": "int64",
          "type": "integer"
        },
        "proprietaireId": {
          "format": "int32",
          "type": "integer"
        },
        "commandes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Commande"
          }
        },
        "montantTotalHTBL": {
          "format": "double",
          "type": "number"
        },
        "remiseBL": {
          "format": "double",
          "type": "number"
        },
        "netHT": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "montantTotalTTCBL": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "BLUpdateDTO": {
      "type": "object",
      "properties": {
        "titreClient": {
          "$ref": "#/definitions/Titre"
        },
        "nomClient": {
          "type": "string"
        },
        "adresseClient": {
          "type": "string"
        },
        "mfClient": {
          "type": "string"
        },
        "gsmClient": {
          "format": "int64",
          "type": "integer"
        },
        "commandes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Commande"
          }
        },
        "remiseBL": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "BonDeLivraison": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "reference": {
          "type": "string"
        },
        "titreClient": {
          "$ref": "#/definitions/Titre"
        },
        "nomClient": {
          "type": "string"
        },
        "adresseClient": {
          "type": "string"
        },
        "mfClient": {
          "type": "string"
        },
        "gsmClient": {
          "format": "int64",
          "type": "integer"
        },
        "proprietaireId": {
          "format": "int32",
          "type": "integer"
        },
        "commandesJson": {
          "type": "string"
        },
        "commandes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Commande"
          }
        },
        "montantTotalHTBL": {
          "format": "double",
          "type": "number"
        },
        "remiseBL": {
          "format": "double",
          "type": "number"
        },
        "netHT": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "montantTotalTTCBL": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "BonDeSortieCreateDTO": {
      "type": "object",
      "properties": {
        "listeFactures": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Facture"
          }
        },
        "matriculeDeVoiture": {
          "type": "string"
        },
        "dateDebutCirculation": {
          "format": "date-time",
          "type": "string"
        },
        "dateFinCirculation": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "BonDeSortieDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "referenceBS": {
          "type": "string"
        },
        "listeFactures": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Facture"
          }
        },
        "matriculeDeVoiture": {
          "type": "string"
        },
        "dateDebutCirculation": {
          "format": "date-time",
          "type": "string"
        },
        "dateFinCirculation": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "BonDeSortieUpdateDTO": {
      "type": "object",
      "properties": {
        "listeFactures": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Facture"
          }
        },
        "matriculeDeVoiture": {
          "type": "string"
        },
        "dateDebutCirculation": {
          "format": "date-time",
          "type": "string"
        },
        "dateFinCirculation": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "Commande": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "referenceCommande": {
          "type": "string"
        },
        "nomVendeur": {
          "type": "string"
        },
        "adresseVendeur": {
          "type": "string"
        },
        "mfVendeur": {
          "type": "string"
        },
        "gsmVendeur": {
          "format": "int64",
          "type": "integer"
        },
        "proprietaireId": {
          "format": "int32",
          "type": "integer"
        },
        "titreAcheteur": {
          "$ref": "#/definitions/Titre"
        },
        "nomAcheteur": {
          "type": "string"
        },
        "adresseAcheteur": {
          "type": "string"
        },
        "mfAcheteur": {
          "type": "string"
        },
        "gsmAcheteur": {
          "format": "int64",
          "type": "integer"
        },
        "listIdsTransactions": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        },
        "montantTotalHT": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "montantTotalTTC": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "Facture": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "referenceFacture": {
          "type": "string"
        },
        "titreClient": {
          "$ref": "#/definitions/Titre"
        },
        "nomClient": {
          "type": "string"
        },
        "adresseClient": {
          "type": "string"
        },
        "mfClient": {
          "type": "string"
        },
        "gsmClient": {
          "format": "int64",
          "type": "integer"
        },
        "proprietaireId": {
          "format": "int32",
          "type": "integer"
        },
        "listeBL": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BonDeLivraison"
          }
        },
        "dateEmission": {
          "format": "date-time",
          "type": "string"
        },
        "dateMiseAJour": {
          "format": "date-time",
          "type": "string"
        },
        "montantHTFacture": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "remise": {
          "format": "double",
          "type": "number"
        },
        "montantTTCFacture": {
          "format": "double",
          "type": "number"
        },
        "listeBLSerialized": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "Titre": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    },
...
  }
}